lambda `L`: explicitly compare with 0 for on-the-curve check
In the `lift_x`-corresponding lambda `L`, check if the
curve equation y**2 == x**3 + 7 holds for the calculated y,
by explicitly comparing with zero, rather than applying the
boolean `not` operator on an integer. The code size stays
the same in terms of characters used (I'm not so sure on the
amount of ink needed for printing the shirt though).

Rationale: crazy folks trying to decipher the shirt,
may gain a few seconds by being slightly less confused
about this part.